+++ dart/test/stack_trace_test.dart
@@ -3,11 +3,11 @@
 // found in the LICENSE file.
 
 import 'package:sentry/src/origin.dart';
+import 'package:sentry/src/platform/mock_platform.dart';
 import 'package:sentry/src/sentry_stack_trace_factory.dart';
 import 'package:stack_trace/stack_trace.dart';
 import 'package:test/test.dart';
 
-import 'mocks/mock_platform_checker.dart';
 import 'test_utils.dart';
 
 void main() {
@@ -284,13 +284,13 @@ isolate_instructions: 10fa27070, vm_instructions: 10fa21e20
       final fixture = Fixture();
 
       // Test for web platform
-      fixture.options.platformChecker = MockPlatformChecker(isWebValue: true);
+      fixture.options.platform = MockPlatform(isWeb: true);
       final webSut = fixture.getSut();
       var webFrame = webSut.encodeStackTraceFrame(frame)!;
       expect(webFrame.platform, 'javascript');
 
       // Test for non-web platform
-      fixture.options.platformChecker = MockPlatformChecker(isWebValue: false);
+      fixture.options.platform = MockPlatform(isWeb: false);
       final nativeFrameBeforeSut = fixture.getSut();
       var nativeFrameBefore =
           nativeFrameBeforeSut.encodeStackTraceFrame(frame)!;
@@ -307,7 +307,7 @@ isolate_instructions: 10fa27070, vm_instructions: 10fa21e20
 }
 
 class Fixture {
-  final options = defaultTestOptions(MockPlatformChecker(isWebValue: false));
+  final options = defaultTestOptions()..platform = MockPlatform(isWeb: false);
 
   SentryStackTraceFactory getSut({
     List<String> inAppIncludes = const [],