+++ flutter/lib/src/web/sentry_js_binding.dart
@@ -1,13 +1,16 @@
 import 'package:flutter/cupertino.dart';
 
 export 'noop_sentry_js_binding.dart'
-    if (dart.library.html) 'html_sentry_js_binding.dart'
     if (dart.library.js_interop) 'web_sentry_js_binding.dart';
 
 abstract class SentryJsBinding {
   void init(Map<String, dynamic> options);
   void close();
   void captureEnvelope(List<Object> envelope);
+  void startSession();
+  Map<dynamic, dynamic>? getSession();
+  void updateSession({int? errors, String? status});
+  void captureSession();
 
   @visibleForTesting
   dynamic getJsOptions();