+++ flutter/test/mocks.mocks.dart
@@ -437,11 +437,16 @@ class MockSentryTracer extends _i1.Mock implements _i3.SentryTracer {
       ) as Map<String, String>);
 
   @override
-  _i11.Future<void> finish({_i2.SpanStatus? status, DateTime? endTimestamp}) =>
+  _i11.Future<void> finish({
+    _i2.SpanStatus? status,
+    DateTime? endTimestamp,
+    _i2.Hint? hint,
+  }) =>
       (super.noSuchMethod(
         Invocation.method(#finish, [], {
           #status: status,
           #endTimestamp: endTimestamp,
+          #hint: hint,
         }),
         returnValue: _i11.Future<void>.value(),
         returnValueForMissingStub: _i11.Future<void>.value(),
@@ -816,11 +821,16 @@ class MockSentrySpan extends _i1.Mock implements _i2.SentrySpan {
       ) as Map<String, dynamic>);
 
   @override
-  _i11.Future<void> finish({_i2.SpanStatus? status, DateTime? endTimestamp}) =>
+  _i11.Future<void> finish({
+    _i2.SpanStatus? status,
+    DateTime? endTimestamp,
+    _i2.Hint? hint,
+  }) =>
       (super.noSuchMethod(
         Invocation.method(#finish, [], {
           #status: status,
           #endTimestamp: endTimestamp,
+          #hint: hint,
         }),
         returnValue: _i11.Future<void>.value(),
         returnValueForMissingStub: _i11.Future<void>.value(),
@@ -1007,12 +1017,13 @@ class MockSentryClient extends _i1.Mock implements _i2.SentryClient {
     _i2.SentryTransaction? transaction, {
     _i2.Scope? scope,
     _i2.SentryTraceContextHeader? traceContext,
+    _i2.Hint? hint,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
           #captureTransaction,
           [transaction],
-          {#scope: scope, #traceContext: traceContext},
+          {#scope: scope, #traceContext: traceContext, #hint: hint},
         ),
         returnValue: _i11.Future<_i2.SentryId>.value(
           _FakeSentryId_5(
@@ -1020,7 +1031,7 @@ class MockSentryClient extends _i1.Mock implements _i2.SentryClient {
             Invocation.method(
               #captureTransaction,
               [transaction],
-              {#scope: scope, #traceContext: traceContext},
+              {#scope: scope, #traceContext: traceContext, #hint: hint},
             ),
           ),
         ),
@@ -1032,14 +1043,6 @@ class MockSentryClient extends _i1.Mock implements _i2.SentryClient {
         Invocation.method(#captureEnvelope, [envelope]),
         returnValue: _i11.Future<_i2.SentryId?>.value(),
       ) as _i11.Future<_i2.SentryId?>);
-
-  @override
-  _i11.Future<void> captureUserFeedback(_i2.SentryUserFeedback? userFeedback) =>
-      (super.noSuchMethod(
-        Invocation.method(#captureUserFeedback, [userFeedback]),
-        returnValue: _i11.Future<void>.value(),
-        returnValueForMissingStub: _i11.Future<void>.value(),
-      ) as _i11.Future<void>);
 
   @override
   _i11.Future<_i2.SentryId> captureFeedback(
@@ -2681,14 +2684,6 @@ class MockHub extends _i1.Mock implements _i2.Hub {
       ) as _i11.Future<_i2.SentryId>);
 
   @override
-  _i11.Future<void> captureUserFeedback(_i2.SentryUserFeedback? userFeedback) =>
-      (super.noSuchMethod(
-        Invocation.method(#captureUserFeedback, [userFeedback]),
-        returnValue: _i11.Future<void>.value(),
-        returnValueForMissingStub: _i11.Future<void>.value(),
-      ) as _i11.Future<void>);
-
-  @override
   _i11.Future<_i2.SentryId> captureFeedback(
     _i2.SentryFeedback? feedback, {
     _i2.Hint? hint,
@@ -2833,12 +2828,13 @@ class MockHub extends _i1.Mock implements _i2.Hub {
   _i11.Future<_i2.SentryId> captureTransaction(
     _i2.SentryTransaction? transaction, {
     _i2.SentryTraceContextHeader? traceContext,
+    _i2.Hint? hint,
   }) =>
       (super.noSuchMethod(
         Invocation.method(
           #captureTransaction,
           [transaction],
-          {#traceContext: traceContext},
+          {#traceContext: traceContext, #hint: hint},
         ),
         returnValue: _i11.Future<_i2.SentryId>.value(
           _FakeSentryId_5(
@@ -2846,7 +2842,7 @@ class MockHub extends _i1.Mock implements _i2.Hub {
             Invocation.method(
               #captureTransaction,
               [transaction],
-              {#traceContext: traceContext},
+              {#traceContext: traceContext, #hint: hint},
             ),
           ),
         ),